New Customer and Organizer Pov w/ added dark mode
For all povs, there is components such as the top navigation bar, sidebar, and footer for all pages and a toggle night mode version of all pages. In the Customer POV, the homepage that should resemble the mockups in Figma. In the Organizer POV, a dashboard page, not in Figma but please consider, and the Bookings Page, resembles the mockups in Figma.